Edit admin access and security improvements
#1

PROGRESS:
- Edit admin access
- Security improvements

+++ web-admin/clear-login-log-and-statistics.php
@@ -3,7 +3,11 @@ session_save_path('/tmp');
 
 include("../php-main/lib.php");
 
-$tables = array(TABLE_LOGIN_LOG, TABLE_LOGIN_STATISTICS);
-lib_database::deleteTables($tables);
+if(lib_get::loginStatus() == STATUS_LOGGED_IN) {
+    if (lib_check::userIsAdmin()) {
+        $tables = array(TABLE_LOGIN_LOG, TABLE_LOGIN_STATISTICS);
+        lib_database::deleteTables($tables);
+    }
+}
 
 lib::redirect(FALSE, NULL, TRUE, NULL);
